fix(wallet-core): extract selectors & thunks to break circular deps
