+++ suite-common/wallet-core/src/index.ts
@@ -1,3 +1,4 @@
+export * from './selectors';
 export * from './accounts/accountsActions';
 export * from './accounts/accountsConstants';
 export * from './accounts/accountsMiddleware';
@@ -46,3 +47,4 @@ export * from './transactions/transactionsThunks';
 export * from './settings/walletSettingsActions';
 export * from './settings/walletSettingsConstants';
 export * from './settings/walletSettingsReducer';
+export * from './settings/walletSettingsThunks';
