+++ suite-common/wallet-core/src/selectors.ts
@@ -1,0 +1,92 @@
+import { createWeakMapSelector } from '@suite-common/redux-utils';
+import { networksCollection } from '@suite-common/wallet-config';
+import { getFailedAccounts } from '@suite-common/wallet-utils';
+import { StaticSessionId } from '@trezor/connect';
+
+import {
+    AccountsRootState,
+    selectAccounts,
+    selectAccountsByDeviceState,
+    selectIsDeviceAccountless,
+} from './accounts/accountsReducer';
+import {
+    DeviceRootState,
+    selectHasOnlyPortfolioDevice,
+    selectSelectedDevice,
+} from './device/deviceReducer';
+import {
+    DiscoveryRootState,
+    selectDiscoveryByDevicePath,
+    selectIsDeviceDiscoveryActive,
+} from './discovery/discoveryReducer';
+import { WalletSettingsRootState, selectEnabledNetworks } from './settings/walletSettingsReducer';
+
+/*
+This file is for selectors that reach into more than one wallet-core reduce
+to prevent circular dependencies between reducers
+*/
+
+const createMemoizedSelector = createWeakMapSelector.withTypes<
+    AccountsRootState & DeviceRootState & DiscoveryRootState
+>();
+
+export const selectNetworksToDiscover = (
+    state: DiscoveryRootState & DeviceRootState & AccountsRootState & WalletSettingsRootState,
+    staticSessionId?: StaticSessionId,
+) => {
+    const enabledNetworks = selectEnabledNetworks(state);
+
+    if (!staticSessionId) {
+        console.log('staticSessionId is not defined, returning full');
+
+        return enabledNetworks;
+    }
+    const device = selectSelectedDevice(state);
+    const discovery = selectDiscoveryByDevicePath(state, device?.path);
+    const okAccounts = selectAccountsByDeviceState(state, staticSessionId);
+    const failedAccounts = getFailedAccounts(staticSessionId, discovery);
+
+    const discoveredNetworks = [
+        ...new Set([...okAccounts, ...failedAccounts].map(account => account.symbol)),
+    ];
+
+    return enabledNetworks.filter(network => !discoveredNetworks.includes(network));
+};
+
+export const selectIsRediscoverNeeded = (
+    state: DiscoveryRootState & DeviceRootState & AccountsRootState & WalletSettingsRootState,
+    staticSessionId?: StaticSessionId,
+) => {
+    if (!staticSessionId) return false;
+
+    const networksToDiscover = selectNetworksToDiscover(state, staticSessionId);
+
+    return networksToDiscover.length > 0;
+};
+
+export const selectAccountsToBeForgotten = (
+    state: DiscoveryRootState & AccountsRootState & WalletSettingsRootState,
+) => {
+    const accounts = selectAccounts(state);
+    const enabledNetworks = selectEnabledNetworks(state);
+    // find disabled networks
+    const disabledNetworks = networksCollection
+        .filter(n => !enabledNetworks.includes(n.symbol) || n.isHidden)
+        .map(n => n.symbol);
+    // find accounts for disabled networks
+    const accountsToRemove = accounts.filter(
+        a => disabledNetworks.includes(a.symbol) && !a.imported,
+    );
+
+    return accountsToRemove;
+};
+
+export const selectIsDiscoveredDeviceAccountless = createMemoizedSelector(
+    [selectIsDeviceAccountless, selectIsDeviceDiscoveryActive],
+    (isAccountless, isDiscoveryActive) => isAccountless && !isDiscoveryActive,
+);
+
+export const selectHasOnlyEmptyPortfolioTracker = createMemoizedSelector(
+    [selectIsDiscoveredDeviceAccountless, selectHasOnlyPortfolioDevice],
+    (isDiscoveredAccountless, hasOnlyPortfolio) => isDiscoveredAccountless && hasOnlyPortfolio,
+);
