+++ suite-common/wallet-core/src/settings/walletSettingsActions.ts
@@ -1,14 +1,11 @@
 import { Dispatch, createAction } from '@reduxjs/toolkit';
 
-import { createThunk } from '@suite-common/redux-utils';
 import { FiatCurrencyCode } from '@suite-common/suite-config';
 import { NetworkSymbol } from '@suite-common/wallet-config';
 import { PROTO } from '@trezor/connect';
 
 import { WALLET_SETTINGS } from './walletSettingsConstants';
-import { selectBitcoinAmountUnit, selectEnabledNetworks } from './walletSettingsReducer';
-import { selectAccountsToBeForgotten } from '../discovery/discoveryReducer';
-import { accountsActions } from '../accounts/accountsActions';
+import { selectBitcoinAmountUnit } from './walletSettingsReducer';
 
 export const setLocalCurrency = createAction(
     WALLET_SETTINGS.SET_LOCAL_CURRENCY,
@@ -55,38 +52,6 @@ export type WalletSettingsAction =
     | ChangeCoinVisibilityAction
     | SetHideBalanceAction
     | SetBitcoinAmountUnitsAction;
-
-export const changeCoinVisibility = createThunk<
-    void,
-    {
-        symbol: NetworkSymbol;
-        shouldBeVisible: boolean;
-    },
-    void
->(WALLET_SETTINGS.CHANGE_COIN_VISIBILITY, ({ symbol, shouldBeVisible }, { dispatch, getState }) => {
-    let enabledNetworks = selectEnabledNetworks(getState());
-    const isAlreadyHidden = enabledNetworks.find(enabledSymbol => enabledSymbol === symbol);
-    if (!shouldBeVisible) {
-        enabledNetworks = enabledNetworks.filter(enabledSymbol => enabledSymbol !== symbol);
-    } else if (!isAlreadyHidden) {
-        enabledNetworks = [...enabledNetworks, symbol];
-    }
-    dispatch(changeNetworks(enabledNetworks));
-
-    const accountsToRemove = selectAccountsToBeForgotten(getState());
-    if (accountsToRemove.length > 0) {
-        dispatch(accountsActions.removeAccount(accountsToRemove));
-    }
-
-    // this seems to be only for analyticsMiddleware
-    dispatch({
-        type: WALLET_SETTINGS.CHANGE_COIN_VISIBILITY,
-        payload: {
-            symbol,
-            shouldBeVisible,
-        },
-    });
-});
 
 export const setDiscreetMode = (toggled: boolean): WalletSettingsAction => ({
     type: WALLET_SETTINGS.SET_HIDE_BALANCE,
